+++ oauth/oauth-core/src/main/groovy/com/junbo/oauth/core/service/impl/ClientServiceImpl.groovy
@@ -4,24 +4,22 @@
  * Copyright (C) 2014 Junbo and/or its affiliates. All rights reserved.
  */
 package com.junbo.oauth.core.service.impl
-
 import com.junbo.authorization.AuthorizeContext
-import com.junbo.oauth.core.exception.AppExceptions
+import com.junbo.common.error.AppCommonErrors
+import com.junbo.oauth.core.exception.AppErrors
 import com.junbo.oauth.core.service.ClientService
-import com.junbo.oauth.core.service.ScopeService
 import com.junbo.oauth.core.service.OAuthTokenService
+import com.junbo.oauth.core.service.ScopeService
 import com.junbo.oauth.core.util.UriUtil
 import com.junbo.oauth.db.exception.DBUpdateConflictException
 import com.junbo.oauth.db.generator.TokenGenerator
 import com.junbo.oauth.db.repo.ClientRepository
 import com.junbo.oauth.spec.model.Client
-import com.junbo.oauth.spec.model.TokenInfo
 import groovy.transform.CompileStatic
 import org.springframework.beans.factory.annotation.Required
 import org.springframework.util.StringUtils
 
 import java.util.regex.Pattern
-
 /**
  * ClientServiceImpl.
  */
@@ -61,7 +59,7 @@ class ClientServiceImpl implements ClientService {
     @Override
     Client saveClient(Client client) {
         if (!AuthorizeContext.hasScopes(CLIENT_REGISTER_SCOPE)) {
-            throw AppExceptions.INSTANCE.insufficientScope().exception()
+            throw AppErrors.INSTANCE.insufficientScope().exception()
         }
 
         validateClient(client)
@@ -83,17 +81,17 @@ class ClientServiceImpl implements ClientService {
     @Override
     Client getClient(String clientId) {
         if (!AuthorizeContext.hasScopes(CLIENT_REGISTER_SCOPE)) {
-            throw AppExceptions.INSTANCE.insufficientScope().exception()
+            throw AppErrors.INSTANCE.insufficientScope().exception()
         }
 
         Client client = clientRepository.getClient(clientId)
 
         if (client == null) {
-            throw AppExceptions.INSTANCE.notExistClient(clientId).exception()
+            throw AppErrors.INSTANCE.notExistClient(clientId).exception()
         }
 
         if (client.ownerUserId != AuthorizeContext.currentUserId) {
-            throw AppExceptions.INSTANCE.notClientOwner().exception()
+            throw AppErrors.INSTANCE.notClientOwner().exception()
         }
 
         return client
@@ -102,13 +100,13 @@ class ClientServiceImpl implements ClientService {
     @Override
     Client getClientInfo(String clientId) {
         if (!AuthorizeContext.hasScopes(CLIENT_INFO_SCOPE)) {
-            throw AppExceptions.INSTANCE.insufficientScope().exception()
+            throw AppErrors.INSTANCE.insufficientScope().exception()
         }
 
         Client client = clientRepository.getClient(clientId)
 
         if (client == null) {
-            throw AppExceptions.INSTANCE.notExistClient(clientId).exception()
+            throw AppErrors.INSTANCE.notExistClient(clientId).exception()
         }
 
         return new Client(
@@ -121,34 +119,42 @@ class ClientServiceImpl implements ClientService {
 
     @Override
     Client updateClient(String clientId, Client client) {
-        if (StringUtils.isEmpty(client.revision)) {
-            throw AppExceptions.INSTANCE.missingRevision().exception()
+        if (!AuthorizeContext.hasScopes(CLIENT_REGISTER_SCOPE)) {
+            throw AppErrors.INSTANCE.insufficientScope().exception()
+        }
+
+        if (StringUtils.isEmpty(client.rev)) {
+            throw AppCommonErrors.INSTANCE.fieldRequired('revision').exception()
         }
 
         Client existingClient = getClient(clientId)
 
-        if (client.revision != existingClient.revision) {
-            throw AppExceptions.INSTANCE.updateConflict().exception()
+        if (client.ownerUserId != AuthorizeContext.currentUserId) {
+            throw AppErrors.INSTANCE.notClientOwner().exception()
         }
 
-        if (client.clientId != null && existingClient.clientId != client.clientId) {
-            throw AppExceptions.INSTANCE.cantUpdateFields('client_id').exception()
+        if (client.rev != existingClient.rev) {
+            throw AppErrors.INSTANCE.updateConflict().exception()
         }
 
-        if (client.clientSecret != null && existingClient.clientSecret != client.clientSecret) {
-            throw AppExceptions.INSTANCE.cantUpdateFields('client_secret').exception()
+        if (client.clientId == null || existingClient.clientId != client.clientId) {
+            throw AppErrors.INSTANCE.cantUpdateFields('client_id').exception()
         }
 
-        if (client.ownerUserId != null && existingClient.ownerUserId != client.ownerUserId) {
-            throw AppExceptions.INSTANCE.cantUpdateFields('owner_user_id').exception()
+        if (client.clientSecret == null || existingClient.clientSecret != client.clientSecret) {
+            throw AppErrors.INSTANCE.cantUpdateFields('client_secret').exception()
         }
 
-        if (client.idTokenIssuer != null && existingClient.idTokenIssuer != client.idTokenIssuer) {
-            throw AppExceptions.INSTANCE.cantUpdateFields('id_token_issuer').exception()
+        if (client.ownerUserId == null || existingClient.ownerUserId != client.ownerUserId) {
+            throw AppErrors.INSTANCE.cantUpdateFields('owner_user_id').exception()
         }
 
-        if (client.needConsent != null && existingClient.needConsent != client.needConsent) {
-            throw AppExceptions.INSTANCE.cantUpdateFields('need_consent').exception()
+        if (client.idTokenIssuer == null || existingClient.idTokenIssuer != client.idTokenIssuer) {
+            throw AppErrors.INSTANCE.cantUpdateFields('id_token_issuer').exception()
+        }
+
+        if (client.needConsent == null || existingClient.needConsent != client.needConsent) {
+            throw AppErrors.INSTANCE.cantUpdateFields('need_consent').exception()
         }
 
         validateClient(client)
@@ -158,29 +164,44 @@ class ClientServiceImpl implements ClientService {
         client.ownerUserId = existingClient.ownerUserId
         client.idTokenIssuer = existingClient.idTokenIssuer
         client.needConsent = existingClient.needConsent
-        client.revision = existingClient.revision
+        client.rev = existingClient.rev
 
         try {
-            return clientRepository.updateClient(client)
+            return clientRepository.updateClient(client, existingClient)
         } catch (DBUpdateConflictException e) {
-            throw AppExceptions.INSTANCE.updateConflict().exception()
+            throw AppErrors.INSTANCE.updateConflict().exception()
         }
     }
 
     @Override
     void deleteClient(String clientId) {
+        if (!AuthorizeContext.hasScopes(CLIENT_REGISTER_SCOPE)) {
+            throw AppErrors.INSTANCE.insufficientScope().exception()
+        }
+
         Client client = getClient(clientId)
+
+        if (client.ownerUserId != AuthorizeContext.currentUserId) {
+            throw AppErrors.INSTANCE.notClientOwner().exception()
+        }
 
         clientRepository.deleteClient(client)
     }
 
     @Override
     Client resetSecret(String clientId) {
+        if (!AuthorizeContext.hasScopes(CLIENT_REGISTER_SCOPE)) {
+            throw AppErrors.INSTANCE.insufficientScope().exception()
+        }
+
         Client client = getClient(clientId)
 
-        client.clientSecret = tokenGenerator.generateClientSecret()
+        if (client.ownerUserId != AuthorizeContext.currentUserId) {
+            throw AppErrors.INSTANCE.notClientOwner().exception()
+        }
 
-        clientRepository.updateClient(client)
+        client.clientSecret = tokenGenerator.generateClientSecret()
+        clientRepository.updateClient(client, client)
         return client
     }
 
@@ -198,8 +219,8 @@ class ClientServiceImpl implements ClientService {
         }
 
         if (invalidScopes != null && !invalidScopes.isEmpty()) {
-            throw AppExceptions.INSTANCE.
-                    invalidScope(StringUtils.collectionToCommaDelimitedString(invalidScopes)).exception()
+            throw AppCommonErrors.INSTANCE.fieldInvalid('scope',
+                    StringUtils.collectionToCommaDelimitedString(invalidScopes)).exception()
         }
 
         if (client.defaultScopes != null) {
@@ -208,7 +229,7 @@ class ClientServiceImpl implements ClientService {
             }
 
             if (invalidScopes != null && !invalidScopes.isEmpty()) {
-                throw AppExceptions.INSTANCE.
+                throw AppErrors.INSTANCE.
                         invalidDefaultScope(StringUtils.collectionToCommaDelimitedString(invalidScopes)).exception()
             }
         }
@@ -217,15 +238,14 @@ class ClientServiceImpl implements ClientService {
     private static void validateRedirectUri(Client client) {
         if (client.redirectUris != null) {
             client.redirectUris.each { String uri ->
-                String escapedUri = uri.replace('*', 'a')
-                if (!UriUtil.isValidUri(escapedUri)) {
-                    throw AppExceptions.INSTANCE.invalidRedirectUri(uri).exception()
+                if (!UriUtil.isValidRedirectUri(uri)) {
+                    throw AppCommonErrors.INSTANCE.fieldInvalid('redirect_uri', uri).exception()
                 }
             }
 
             if (client.defaultRedirectUri != null) {
                 if (client.defaultRedirectUri.contains('*')) {
-                    throw AppExceptions.INSTANCE.invalidDefaultRedirectUri(client.defaultRedirectUri,
+                    throw AppErrors.INSTANCE.invalidDefaultRedirectUri(client.defaultRedirectUri,
                             'the default redirect uri cannot include wildcard').exception()
                 }
 
@@ -234,7 +254,7 @@ class ClientServiceImpl implements ClientService {
                 }
 
                 if (!valid) {
-                    throw AppExceptions.INSTANCE.invalidDefaultRedirectUri(client.defaultRedirectUri,
+                    throw AppErrors.INSTANCE.invalidDefaultRedirectUri(client.defaultRedirectUri,
                             'the default redirect uri must match at least one redirect uri').exception()
                 }
             }
@@ -244,15 +264,14 @@ class ClientServiceImpl implements ClientService {
     private static void validateLogoutRedirectUri(Client client) {
         if (client.logoutRedirectUris != null) {
             client.logoutRedirectUris.each { String uri ->
-                String escapedUri = uri.replace('*', 'a')
-                if (!UriUtil.isValidUri(escapedUri)) {
-                    throw AppExceptions.INSTANCE.invalidLogoutRedirectUri(uri).exception()
+                if (!UriUtil.isValidRedirectUri(uri)) {
+                    throw AppErrors.INSTANCE.invalidLogoutRedirectUri(uri).exception()
                 }
             }
 
             if (client.defaultLogoutRedirectUri != null) {
                 if (client.defaultLogoutRedirectUri.contains('*')) {
-                    throw AppExceptions.INSTANCE.invalidDefaultLogoutRedirectUri(client.defaultLogoutRedirectUri,
+                    throw AppErrors.INSTANCE.invalidDefaultLogoutRedirectUri(client.defaultLogoutRedirectUri,
                             'the default logout redirect uri cannot include wildcard').exception()
                 }
 
@@ -261,7 +280,7 @@ class ClientServiceImpl implements ClientService {
                 }
 
                 if (!valid) {
-                    throw AppExceptions.INSTANCE.invalidDefaultLogoutRedirectUri(client.defaultLogoutRedirectUri,
+                    throw AppErrors.INSTANCE.invalidDefaultLogoutRedirectUri(client.defaultLogoutRedirectUri,
                             'the default logout redirect uri must match at least one logout redirect uri').exception()
                 }
             }
@@ -270,7 +289,7 @@ class ClientServiceImpl implements ClientService {
 
     private static void validateLogoUri(Client client) {
         if (client.logoUri != null && !UriUtil.isValidUri(client.logoUri)) {
-            throw AppExceptions.INSTANCE.invalidLogoUri(client.logoUri).exception()
+            throw AppErrors.INSTANCE.invalidLogoUri(client.logoUri).exception()
         }
     }
 
@@ -278,7 +297,7 @@ class ClientServiceImpl implements ClientService {
         if (client.contacts != null) {
             client.contacts.each { String contact ->
                 if (!EMAIL_PATTERN.matcher(contact)) {
-                    throw AppExceptions.INSTANCE.invalidContacts(contact).exception()
+                    throw AppErrors.INSTANCE.invalidContacts(contact).exception()
                 }
             }
         }
